+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Changelog
 * added: indicate remote rigs in `_iblrig_taskSettings.raw.json`
 * added: `ibl-photometry` dependency
 * changed: reduce traceback size in UDP message
+* changed: hardware validation - more solid way to estimate TTL rates
+* changed: clean-up after failed session (terminate Bonsai processes and stop state machine)
 
 -------------------------------